refactor response class out
